Created Enums.ts
